// src/app/models/i-rules-trend-context.ts
export interface IRulesTrendContext {
        businessName: string;
        profileNum: number;
        baseTableName: string;
        ruleDimension: string;
        ruleCategory: string;
        ruleType: string;
        ruleName: string;
        ruleDescription: string;
        ruleFilter: string;
        ruleSuccessCondition: string;
}